Above capacity must be above zero to delete

+++ app/controllers/images_controller.rb
@@ -33,9 +33,11 @@ class ImagesController < ApplicationController
     free_capacity = 10000 - parallel_uploads_max
     actual_row_count = Image.count
     above_capacity = actual_row_count - free_capacity
-    Image.limit(above_capacity).find_each(batch_size: 50) do |img|
-      Rails.logger.warn("[DELETE] #{img.id}: \"#{img.token}\"")
-      img.destroy!
+    if above_capacity > 0
+      Image.limit(above_capacity).find_each(batch_size: 50) do |img|
+        Rails.logger.warn("[DELETE] #{img.id}: \"#{img.token}\"")
+        img.destroy!
+      end
     end
 
     respond_to do |format|
